Move the factory test mode dialog out of the "language" panel
This dialog was in the "language" panel just because, being the first
page that was shown during the FBE, that happened to be the place
where it was opened from. However, that won't longer be the case now
that we can have "branded" welcome pages, so we need to move this to a
more generic location, that we can potentially reuse from any page.

This commit moves the code from gis-language-page.c to gis-page-util.c, as
well as the UI definitions for the factory mode to its own .ui file.

https://phabricator.endlessm.com/T15834
