+++ gnome-initial-setup/gis-page-util.gresource.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<gresources>
+  <gresource prefix="/org/gnome/initial-setup">
+    <file preprocess="xml-stripblanks">gis-page-util.ui</file>
+  </gresource>
+</gresources>
